Migliorie alla classe ShowData
- Creato il file FlatLaf.properties per definire le proprietà comuni ad ogni tema
- Rimosso il componente ModaList
- Aggiunte 3 proprietà per lo stile della JList di ShowData
