+++ src/main/resources/moda/passwordmanager/FlatDarkLaf.properties
@@ -1,5 +1,5 @@
-TabbedPane.showTabSeparators = true
-TabbedPane.underlineColor = #000000
-TabbedPane.inactiveUnderlineColor = #000000
 Moda.Sidebar.background = #000000
 Moda.Sidebar.Title.foreground = #ffffff
+Moda.ShowData.DataList.selectionBackground = #000000
+Moda.ShowData.DataList.evenBackground = #ffffff
+Moda.ShowData.DataList.oddBackground = #f1f1f1
