+++ src/main/resources/moda/passwordmanager/FlatLaf.properties
@@ -1,0 +1,3 @@
+TabbedPane.showTabSeparators = true
+TabbedPane.underlineColor = #000000
+TabbedPane.inactiveUnderlineColor = #000000
